Toolbar actions - don't import from a pack
a webpack pack is the root of all js (one of them),
importing a pack from a different one means importing the whole pack twice,
while importing from the same one will error

so anything we need to export, shouldn't live in a pack

(but one test needs to import the pack to enable the `subscribeToRx` side effect)
