+++ app/javascript/packs/toolbar-actions-common.js
@@ -1,16 +1,6 @@
 import { onDelete } from '../toolbar-actions/delete';
 import { onCustomAction } from '../toolbar-actions/custom-action';
-
-function transformResource(resource) {
-  return ({ id: resource });
-}
-
-export default function getGridChecks() {
-  if (ManageIQ.gridChecks.length === 0) {
-    return [ManageIQ.record.recordId].map(transformResource);
-  }
-  return ManageIQ.gridChecks.map(transformResource);
-}
+import { getGridChecks } from '../toolbar-actions/util.js';
 
 function callMapperFunction(eventMapper, event) {
   return Object.prototype.hasOwnProperty.call(eventMapper, event.type)
