+++ app/javascript/toolbar-actions/util.js
@@ -1,0 +1,10 @@
+function transformResource(resource) {
+  return ({ id: resource });
+}
+
+export function getGridChecks() {
+  if (ManageIQ.gridChecks.length === 0) {
+    return [ManageIQ.record.recordId].map(transformResource);
+  }
+  return ManageIQ.gridChecks.map(transformResource);
+}
